% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bacalini_sampleSheet}
\alias{Bacalini_sampleSheet}
\title{Bacalini sample sheet}
\format{
A data frame with 87 rows and columns:
\describe{
   \item{gsm}{Unique identifier for every sample.}
   \item{Group}{A group to which the sample belongs. One of DSP, DSM or DSS.}
}
}
\source{
Original sample sheet was downloaded from the Gene Expression Omnibus (GEO)
data repository as part of the data set with accession number GSE52588
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52588}.
}
\usage{
Bacalini_sampleSheet
}
\description{
Description file for samples in Bacalini data set \insertCite{bacalini2015identification}{meNet}.
87 samples have a family-based structure with a Down syndrome patient, a mother
and one unaffected sibling coming from each of 29 included families. Thus, each
sample belongs to one of three groups: Down syndrome patients (DSP), Down
syndrome mothers (DSM) or Down syndrome siblings (DSS). `meNet::Bacalini_sampleSheet`
is a subset of the original sample sheet which includes more details for every
samples.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
