% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxed_heatmap.R
\name{boxed_heatmap}
\alias{boxed_heatmap}
\title{Heat map without the use of `layout` function}
\usage{
boxed_heatmap(
  cor_matrix,
  clrs_list = c("blue", "white", "red"),
  title = "",
  color_title = "black",
  cex_title = 1.3,
  line_title = 0.5,
  margin_add = 1.5,
  box_width = 1
)
}
\arguments{
\item{cor_matrix}{Correlation matrix.}

\item{clrs_list}{A list of colors passed to `graphics::colorRampPalette`. 
If three colors are given, first corresponds to negative correlation, second 
to zero correlation and third to positive correlation.
Defaults to `c("blue", "white", "red")`.}

\item{title}{Title for the heat map. Passed to the `graphics::title` function.}

\item{color_title}{Color of the title. Passed to the `graphics::title` function.}

\item{cex_title}{Size of the title. Passed to the `graphics::title` function.}

\item{line_title}{Line at which title is added to the plot. Passed to the `graphics::title`
function.}

\item{margin_add}{Additional size for the margins. `c(0.5,0.5,0.5,0.5)+margin_add`
is passed as the `mar` parameter to `graphics::par`. Defaults to 1.5.}

\item{box_width}{Width of the box around the heat map. 
Passed to the `graphics::box` function.}
}
\description{
Function which draws heat map with box around it and without any legend.
The function calls `image` and because of its simplicity it doesn't use `layout` so it can be used
inside the `layout` function together with other plots.
}
