% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_resamplingStats.R
\name{cor_resamplingStats}
\alias{cor_resamplingStats}
\title{Mean and standard deviation of correlation}
\usage{
cor_resamplingStats(
  data,
  size_of_subsample,
  replace = FALSE,
  n_repetitions = 100,
  zero_precisionC = 1e-06
)
}
\arguments{
\item{data}{Data matrix or data frame with samples in rows and variables in columns.}

\item{size_of_subsample}{Size of subsample taken in every repetition.}

\item{replace}{Defines whether the resampling is done with replacement. Defaults to `FALSE`.}

\item{n_repetitions}{Number of repetitions of subsampling. For large `data` matrices, large
number of repetitions can take a long time.}

\item{zero_precisionC}{The zero precision passed to the C++ function. All numbers with absolute
value smaller than `zero_precisionC` will be set to zero. Defaults to 10^(-6).}
}
\value{
A data frame with names of variables in first two columns. In third columns is mean
and in the fourth column is standard deviation for the given pair of variables. In the rows, 
all pairwise correlations are listed.
}
\description{
For the given data matrix, the mean and standard deviation of correlation are 
calculated for every pair of variables. The mean and standard deviation are calculated
based on correlation values obtained after random subsampling.
}
\details{
For `n_repetitions` times, a random subsample is taken based on parameters 
`size_of_subsample` and `replace`. Then for every pair of variables a correlation is calculated
for a random subsample. At the end of repetitions, the mean and standard deviation of correlation
are calculated for every pair of variables.
}
