% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpG_anno450K}
\alias{CpG_anno450K}
\title{Description of HM450K CpG sites}
\format{
A data frame with 485512 rows and variables:
\describe{
   \item{IlmnID}{Unique CpG identifier from the Illumina CG database.}
   \item{CHR}{Chromosome containing the CpG.}
   \item{MAPINFO}{Chromosomal coordinates of the CpG.}
   \item{Strand}{The Forward (F) or Reverse (R) designation of the Design 
   Strand.}
   \item{UCSC_RefGene_Name}{Target gene name(s), from the UCSC database. 
   *Note: multiple listings of the same gene name indicate splice variants.}
   \item{UCSC_RefGene_Accession}{The UCSC accession number(s) of the target 
   transcript(s). Accession numbers are given in the same order as the target
   gene transcripts.}
   \item{UCSC_RefGene_Group}{Gene region feature category describing the CpG 
   position, from UCSC. Features listed in the same order as the target gene 
   transcripts.}
   \item{UCSC_CpG_Islands_Name}{Chromosomal coordinates of the CpG island 
   from UCSC}
   \item{Relation_to_UCSC_CpG_Island}{The location of the CpG relative to the
   CpG island.}
}
}
\source{
\url{https://support.illumina.com/downloads/humanmethylation450_15017482_v1-2_product_files.html}
}
\usage{
CpG_anno450K
}
\description{
Description file for CpGs  measured by Illumina 450K microarray.
It is Illumina `“HumanMethylation450 15017482 v1-2.csv”` manifest file with
reduced number of columns.
}
\keyword{datasets}
