% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpG_genes}
\alias{CpG_genes}
\title{Genes associated with HM450K CpG sites}
\format{
A data frame 687137 rows and 6 variables:
\describe{
   \item{IlmnID}{Unique CpG identifier from Illumina CG database.}
   \item{CHR}{Chromosome containing the CpG.}
   \item{MAPINFO}{Chromosomal coordinate of the CpG.}
   \item{UCSC_RefGene_Name}{Target gene name, from the UCSC database.}
   \item{UCSC_RefGene_Accession}{The UCSC accession number of the target transcript.}
   \item{UCSC_RefGene_Accession}{Gene region describing the CpG position.}
}
}
\usage{
CpG_genes
}
\description{
Data frame based on the `CpG_anno450K`, Illumina Infinium HumanMethylation450
manifest file, describing the relationship between CpG sites and
genes/transcripts. If CpG is associated with multiple genes/transcripts,
it is present in multiple rows.
}
\keyword{datasets}
