% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_CpG.R
\name{find_CpG}
\alias{find_CpG}
\title{CpG coordinates on a chromosome}
\usage{
find_CpG(chromosome, min_coord = NULL, max_coord = NULL)
}
\arguments{
\item{chromosome}{Name of a human chromosome in the format `"chr?"`, where `"?"`
must either be a number from `1` to `22` or `"X"` or `"Y"`.}

\item{min_coord}{Minimum returned coordinate.}

\item{max_coord}{Maximum returned coordinate.}
}
\value{
Vector of CpG coordinates.
}
\description{
For a given chromosome, returns positions of all "CG" 
dinucleotides (CpG sites) based on a human genome assembly hg19. If minimum
or maximum position is given, only CpGs in the given range are returned.
}
\details{
Returned position is a position of "C". The same convention is used
by Illumina Infinium HM450K manifest file.
Function is based on R package `"BSgenome.Hsapiens.UCSC.hg19"`
\insertCite{team2020bsgenome}{meNet}.
}
\references{
\insertAllCited{}
}
