% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Johansson_residuals}
\alias{Johansson_residuals}
\title{Johansson HM450K methylation data}
\format{
A data frame with 1000 rows and 728 columns. Row names are unique CpG
identifiers from the Illumina CG database.
}
\source{
Original data set was downloaded from the Gene Expression Omnibus (GEO)
data repository as a data set with accession number GSE87571
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87571}.
}
\usage{
Johansson_residuals
}
\description{
Methylation values based on the data set obtained by Johansson et al.
\insertCite{johansson2013continuous}{meNet}. Methylation beta values for 728
healthy samples were obtained with Illumina HM450K microarray. In the step of
data preparation and preprocessing, problematic CpGs were removed and the
original beta values were adjusted for age, sex and cell counts.
`meNet::Johansson_residuals` data frame contains the residuals of the 
preprocessing for 1000 CpG sites from chromosome 21.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
