% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{max_normalization}
\alias{max_normalization}
\title{Maximum normalization}
\usage{
max_normalization(x = numeric(0))
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Normalized numeric vector.
}
\description{
Performs the maximum normalization dividing every element of the 
vector with the maximum absolute value.
}
\examples{
{
max_normalization(c(1,2,3,4))
}

}
