% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meMultiplex_to_multinet.R
\name{meMultiplex_to_multinet}
\alias{meMultiplex_to_multinet}
\title{Multiplex: `meNet` file to `multinet` object}
\usage{
meMultiplex_to_multinet(input_file)
}
\arguments{
\item{input_file}{File with the `multinet` structure of the multiplex.}
}
\value{
Multiplex network as `multinet` object.
}
\description{
Converts file with multiplex structure written in `multinet`-style
to `multinet` object. The file can be created with `meNet::meMultiplex`
function with the output type set to `"multinet"`. This allows the use
of different `multinet` function for multilayer networks 
\insertCite{multinet}{meNet}.
}
\references{
\insertAllCited{}
}
