\name{meNet-package}
\alias{meNet-package}
\alias{meNet}
\docType{package}
\title{
\packageTitle{meNet}
}
\description{
\packageDescription{meNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{meNet}
\packageIndices{meNet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{meNet}

Maintainer: \packageMaintainer{meNet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
