% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meNet_singleGene.R
\name{meNet_singleGene}
\alias{meNet_singleGene}
\title{CpG network for a single gene}
\usage{
meNet_singleGene(
  gene_name,
  cor_matrix = NULL,
  data = NULL,
  link_method = "twoLyr_clust",
  weighted = TRUE,
  cor_normalization_fun = max_normalization,
  dist_normalization_fun = neg_max_normalization,
  cor_threshold = 0.2,
  neg_cor_threshold = NULL,
  cor_stDev = NULL,
  cor_alpha = NULL,
  n_repetitions = 1000,
  alternative = "two_sided",
  infomap_call = "infomap",
  folder = "./meNet/",
  file_basename = "meNet_gene_infomap",
  relaxation_rate = 0.15,
  cgGene_meta = data("CpG_genes", package = "meNet"),
  cgGene_meta_cols = list(cg_id = "IlmnID", cg_coord = "MAPINFO", gene_id =
    "UCSC_RefGene_Name", gene_region = "UCSC_RefGene_Group"),
  gene_regions = c("Promoter", "Body", "3'UTR"),
  check_matrices = TRUE,
  delete_files = FALSE
)
}
\arguments{
\item{gene_name}{}
}
\value{

}
\description{
Builds a CpG network for a single gene with (optional) weights 
representing chromosomal distances between CpGs. 
Edge significance is determined similarly as in singleCGI_meNet with an 
additional clustering method “twoLyr_clust_wRegion” which additionally 
incorporates the gene-region-information (e.g. “Promoter”, “Body”) 
in the distance layer prior to the Infomap multiplex clustering.
}
\details{
CpG network: edges based on gene affiliation
}
