% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N_ranges}
\alias{N_ranges}
\title{Unsequences genome}
\format{
A data frame with 362 rows and 4 variables:
\describe{
   \item{`chr`}{Chromosome containing unsequenced interval.}
   \item{`start`}{Chromosomal coordinate of the first nuclotide within the interval.}
   \item{`end`}{Chromosomal coordinate of the last nuclotide within the interval.}
   \item{`length`}{Length of the interval, expressed in base pairs.}
}.
}
\usage{
N_ranges
}
\description{
Data frame with unsequenced genomic regions for human genome assembly hg19.
Unsequences nucleotides are expressed as intervals since the unsequenced
regions span over larger portions of the genome.
}
\keyword{datasets}
