% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCI_meCpG.R
\name{plotCI_meCpG}
\alias{plotCI_meCpG}
\title{Beta value error plot with CpG coordinates}
\usage{
plotCI_meCpG(
  beta_values,
  sample_groups,
  cg_list = NULL,
  chr = NULL,
  first_coord = NULL,
  last_coord = NULL,
  add_cg = FALSE,
  add_genes = FALSE,
  add_exons = FALSE
)
}
\description{
Beta value error plot with CpG coordinates
}
\details{
Plots error bars and confidence intervals for beta values of given 
CpGs. If list of CpGs is given, they are plotted equidistantly. 
If a range of coordinates and a chromosome are given, then all CpGs inside 
the range are plotted and positioned on x-axis based on their chromosomal 
coordinate. If the latter case, positions of unmeasured CpGs inside the 
region can be added to the plot.
}
