% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC_genes}
\alias{UCSC_genes}
\title{UCSC transcripts}
\format{
A data frame with 78288 rows and variables:
\describe{
   \item{name}{Name of gene/transcript.}
   \item{chrom}{Chromosome containing the gene.}
   \item{strand}{Genome strand on which the gene is located.}
   \item{txStart}{Transcription start position (or end position for minus strand).}
   \item{txEnd}{Transcription end position (or start position for minus strand).}
   \item{cdsStart}{Coding region start position (or end position for minus strand).}
   \item{cdsEnd}{Coding region end position (or start position for minus strand).}
   \item{exonCount}{Number of exons.}
   \item{exonStarts}{Exon start positions (or end positions for minus strand).}
   \item{exonEnds}{Exon end position (or start positions for minus strand).}
}
}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
UCSC_genes
}
\description{
Data frame with description of both protein-coding and non-protein-coding 
transcripts including positions of exons. Transcript name starts with
`"NM"` for protein-coding and with `"NR"` for non-protein coding transcripts.
It is downloaded as `“UCSC RefSeq (refGene)”` table from the UCSC Table
Browser for assembly `“Feb. 2009 (GRCh37/hg19)”` and group 
`“Genes and Gene Predictions”`.
}
\keyword{datasets}
