% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpG_anno450K}
\alias{CpG_anno450K}
\title{Illumina Infinium HumanMethylation450 manifest file}
\format{
A data frame with CpG sites in the rows and their description in the
columns.
}
\source{
\url{https://support.illumina.com/downloads/humanmethylation450_15017482_v1-2_product_files.html}
}
\usage{
CpG_anno450K
}
\description{
Illumina “HumanMethylation450 15017482 v1-2.csv” manifest file with the
description of CpG sites measured by Illumina 450K microarray.
}
\keyword{datasets}
