% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Johansson_residuals}
\alias{Johansson_residuals}
\title{Johansson HM450K methylation data}
\format{
A data frame with 3536 rows and 728 columns. Row names are unique CpG
identifiers from the Illumina CG database.
}
\source{
Original data set was downloaded from the Gene Expression Omnibus (GEO)
data repository as a data set with accession number GSE87571
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE87571}.
}
\usage{
Johansson_residuals
}
\description{
Methylation values based on the data set obtained by Johansson et al.
\insertCite{johansson2013continuous}{meNet}. Methylation beta values for 728
healthy samples were obtained with Illumina HM450K microarray. In the step of
data preparation and preprocessing, problematic CpGs were removed and the
original beta values were adjusted for age, sex and cell counts.
`meNet::Johansson_residuals` data frame contains the residuals of the 
preprocessing analysis for chromosome 21.
}
\details{
Some CpG sites from chromosome 21 are not included. They may be 
omitted in the original data set or removed during the preprocessing. 
Moreover, CpG sites are the same as in the  `meNet::Bacalini_betaValues`
data set and thus only the intersection of measured CpGs remains.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
