% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC_CGI}
\alias{UCSC_CGI}
\title{Data frame with all UCSC islands in rows and columns: CpG_island_name, chr, start, end, 
length, strand and UCSC_CpG_island_name. 
The last column matches the CGI names used in Infinium HumanMethylation450 description file. 
UCSC CpG islands}
\format{
A data frame with 27718 rows and 6 variables:
\describe{
   \item{CpG_island_name}{Name of the CGI.}
   \item{chr}{Chromosome containing the GCI.}
   \item{start}{Chromosomal coordinate of the start of the CGI.}
   \item{end}{Chromosomal coordinate of the end of the CGI.}
   \item{length}{Length of the CGI, in base pairs.}
   \item{UCSC_CpG_Islands_Name}{Name of the CGI which matches 
   `UCSC_CpG_Islands_Name` used in Illumina 450K manifest file,
   or `meNet::CpG_anno450K` dataset.}
}
}
\usage{
UCSC_CGI
}
\description{
Data frame with all UCSC CpG islands (CGI).
}
\keyword{datasets}
